requirements.txt
Django==1.9
djangorestframework==3.3.2
httpie==0.9.3
mysqlclient==1.3.7
psycopg2==2.6.1
Pygments==2.0.2
requests==2.9.1
wheel==0.24.0
